Revert "Restructured questions to linked lists"
