+++ database/generate-qs.js
@@ -1,4 +1,4 @@
-const QuestionModel = require('./questionModel.js');
+const Question = require('./question-model.js');
 
 // To enter Mongo shell: mongo --shell
 
@@ -122,16 +122,14 @@ const resetStarterQs = () => {
   refreshDB(questions);
 };
 
-async function refreshDB(questions) {
-  await QuestionModel.Question.deleteMany({}, (err, res) => {
-    console.log('refreshed questions: ', err || res);
+function refreshDB(questions) {
+  Question.deleteMany({}, (err, res) => {
+    console.log('refreshed database: ', err ? err : res);
+    // TODO: Use Model.bulkWrite instead of Model.create
+    Question.create(questions, (err) => {
+      if (err) console.log('mongoose create', err);
+    });
   });
-  await QuestionModel.Position.deleteMany({}, (err, res) => {
-    console.log('refreshed positions: ', err || res);
-  });
-  for (let i = 0; i < questions.length; i += 1) {
-    await QuestionModel.createQuestion(questions[i]);
-  }
 }
 
 module.exports = resetStarterQs;
