+++ server/server.js
@@ -3,7 +3,7 @@ const express = require('express');
 const path = require('path');
 const bodyParser = require('body-parser');
 const questionController = require('./questionController.js');
-// const resetQs = require('../database/generateQs.js');
+// const resetQs = require('../database/generate-qs.js');
 // resetQs();
 
 const app = express();
